Refactor: Profile부분 폴더 만들어서 기능 세분화
- 사유: 너무 activity, fragment, dialog가 많아서 가독성 떨어졌음
