+++ app/src/main/java/com/example/geeksasaeng/Profile/MyInfoUpdate/DialogProfileCurrentPwdChecking.kt
@@ -1,10 +1,9 @@
-package com.example.geeksasaeng.Profile
+package com.example.geeksasaeng.Profile.MyInfoUpdate
 
 import android.content.Intent
 import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
