+++ app/src/main/java/com/example/geeksasaeng/Profile/MyInfoUpdate/ProfileMyInfoUpdatePwdActivity.kt
@@ -1,7 +1,6 @@
-package com.example.geeksasaeng.Profile
+package com.example.geeksasaeng.Profile.MyInfoUpdate
 
 import android.content.Intent
-import android.os.Bundle
 import android.text.Editable
 import android.text.TextWatcher
 import android.util.Log
