+++ app/src/main/java/com/example/geeksasaeng/Profile/ProfileCard/DialogProfileDetail.kt
@@ -1,4 +1,4 @@
-package com.example.geeksasaeng.Profile
+package com.example.geeksasaeng.Profile.ProfileCard
 
 import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
